Added stuyff

+++ bot/system/interactions.js
@@ -53,17 +53,65 @@ async function loadCommands(client) {
     }
   }
 
-  // Find all commands that are registered on Discord but not in the bot
-  client.application.commands.fetch().then((commands) => {
-    commands.forEach((command) => {
-      if (!client.slashcommands.has(command.name)) {
-        client.log(`Deleting the /${command.name} command.`);
-        command.delete();
-      }
-    });
-  });
-
   client.log(`Loaded all commands.`);
+}
+
+// Context loader
+async function loadContexts(client) {
+  // Find all folders in the contexts directory
+  const commandFolders = fs
+    .readdirSync("./interactions/contexts")
+    .filter((file) =>
+      fs.statSync(path.join("./interactions/contexts", file)).isDirectory()
+    );
+
+  // Register all contexts in the folders
+  for (const folder of commandFolders) {
+    const commandFiles = fs
+      .readdirSync(`./interactions/contexts/${folder}`)
+      .filter((file) => file.endsWith(".js"));
+
+    // Register all slash contexts
+    for (const file of commandFiles) {
+      const command = require(`../interactions/contexts/${folder}/${file}`);
+      if (command.menu) {
+        // Get current slash contexts
+        const contexts = await client.application.commands.fetch();
+
+        // Find a command with the same name in the contexts
+        const existingContext = contexts.find(
+          (cmd) => cmd.name === command.menu.name
+        );
+
+        // If the command doesn't exist, create it
+        if (!existingContext) {
+          client.log(`Created ${command.menu.name} context.`);
+          client.application.commands.create(command.menu);
+        }
+
+        client.contexts.set(command.menu.name, command);
+      }
+    }
+  }
+
+  client.log(`Loaded all contexts.`);
+}
+
+// Delete old contexts and commands
+async function deleteInteractions(client) {
+  // Get current slash commands
+  const commands = await client.application.commands.fetch();
+
+  // Find all commands that are registered on Discord but not in the bot
+  commands.forEach((command) => {
+    if (
+      !client.slashcommands.has(command.name) &&
+      !client.contexts.has(command.name)
+    ) {
+      client.log(`Deleting ${command.name} interaction.`);
+      command.delete();
+    }
+  });
 }
 
 // Button loader
@@ -95,20 +143,18 @@ async function loadButtons(client) {
 async function loadEvents(client) {
   // Find all folders in the events directory
   const eventFolders = await fs
-    .readdirSync("./interactions/events")
-    .filter((file) =>
-      fs.statSync(path.join("./interactions/events", file)).isDirectory()
-    );
+    .readdirSync("./events")
+    .filter((file) => fs.statSync(path.join("./events", file)).isDirectory());
 
   // Register all events in the folders
   for (const folder of eventFolders) {
     const eventFiles = fs
-      .readdirSync(`./interactions/events/${folder}`)
+      .readdirSync(`./events/${folder}`)
       .filter((file) => file.endsWith(".js"));
 
     // Register all events
     for (const file of eventFiles) {
-      const event = require(`../interactions/events/${folder}/${file}`);
+      const event = require(`../events/${folder}/${file}`);
       event.once
         ? client.once(event.name, (...args) => event.execute(...args, client))
         : client.on(event.name, (...args) => event.execute(...args, client));
@@ -118,10 +164,40 @@ async function loadEvents(client) {
   client.log(`Loaded all events.`);
 }
 
+// Modal loader
+async function loadModals(client) {
+  // Find all folders in the modals directory
+  const modalFolders = fs
+    .readdirSync("./interactions/modals")
+    .filter((file) =>
+      fs.statSync(path.join("./interactions/modals", file)).isDirectory()
+    );
+
+  // Register all modals in the folders
+  for (const folder of modalFolders) {
+    const modalFiles = fs
+      .readdirSync(`./interactions/modals/${folder}`)
+      .filter((file) => file.endsWith(".js"));
+
+    // Register all modals
+    for (const file of modalFiles) {
+      const modal = require(`../interactions/modals/${folder}/${file}`);
+      client.modals.set(modal.name, modal);
+    }
+  }
+
+  client.log(`Loaded all modals.`);
+}
+
+// Interaction loader
 async function loadInteractions(client) {
   await loadCommands(client);
+  await loadContexts(client);
   await loadButtons(client);
   await loadEvents(client);
+  await loadModals(client);
+
+  await deleteInteractions(client);
 }
 
 module.exports = {
